refactor(Hand): AtomicInteger 제거
